Completed methods for send bucks

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -3,9 +3,10 @@ package com.techelevator.tenmo.model;
 import java.math.BigDecimal;
 
 public class Transfer {
+
     private Long transferId;
-    private Long transferTypeId;
-    private Long transferStatusId;
+    private int transferTypeId;
+    private int transferStatusId;
     private Long accountFrom;
     private Long accountTo;
     private BigDecimal amount;
@@ -13,7 +14,7 @@ public class Transfer {
     public Transfer() {
     }
 
-    public Transfer(Long transferId, Long transferTypeId, Long transferStatusId, Long accountFrom, Long accountTo, BigDecimal amount) {
+    public Transfer(Long transferId, int transferTypeId, int transferStatusId, Long accountFrom, Long accountTo, BigDecimal amount) {
         this.transferId = transferId;
         this.transferTypeId = transferTypeId;
         this.transferStatusId = transferStatusId;
@@ -30,19 +31,19 @@ public class Transfer {
         this.transferId = transferId;
     }
 
-    public Long getTransferTypeId() {
+    public int getTransferTypeId() {
         return transferTypeId;
     }
 
-    public void setTransferTypeId(Long transferTypeId) {
+    public void setTransferTypeId(int transferTypeId) {
         this.transferTypeId = transferTypeId;
     }
 
-    public Long getTransferStatusId() {
+    public int getTransferStatusId() {
         return transferStatusId;
     }
 
-    public void setTransferStatusId(Long transferStatusId) {
+    public void setTransferStatusId(int transferStatusId) {
         this.transferStatusId = transferStatusId;
     }
 
@@ -69,4 +70,5 @@ public class Transfer {
     public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
+
 }
